pyproject.toml
# GCO (Global Capacity Orchestrator on AWS) - Python Project Configuration

[build-system]
requires = ["setuptools>=77.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["cli", "cli.*", "gco", "gco.*"]

[tool.setuptools.dynamic]
version = {attr = "gco._version.__version__"}

[project]
name = "gco-cli"
dynamic = ["version"]
description = "CLI for managing GCO multi-region EKS clusters"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT-0"
license-files = ["LICENSE"]
authors = [
    {name = "Jacob Mevorach (@Jmevorach)"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "boto3==1.42.96",
    "click==8.3.3",
    "requests==2.33.1",
    "pyyaml==6.0.3",
    "constructs==10.6.0",
    "kubernetes==35.0.0",
    "fastapi==0.136.1",
    "uvicorn==0.46.0",
    "pydantic==2.13.3",
    "prometheus-client==0.25.0",
    "httpx==0.28.1",
    "urllib3>=2.6.3",
]

[project.optional-dependencies]
cdk = [
    "aws-cdk-lib==2.248.0",
    "aws-pdk==0.26.15",
    "cdk-nag==2.37.55",
    "constructs==10.6.0",
]
inference-monitor = [
    "boto3==1.42.96",
    "kubernetes==35.0.0",
    "pyyaml==6.0.3",
]
mcp = [
    "fastmcp>=3.2.0",
]
lint = [
    "black==26.3.1",
    "ruff==0.15.7",
    "isort==8.0.1",
    "yamllint==1.38.0",
    "flake8==7.3.0",
    "flake8-bugbear==25.11.29",
    "flake8-comprehensions==3.16.0",
]
typecheck = [
    "mypy==1.19.1",
    "types-requests==2.32.4.20260107",
    "types-PyYAML==6.0.12.20250915",
]
test = [
    "pytest==9.0.3",
    "pytest-cov==7.1.0",
    "pytest-mock==3.15.1",
    "pytest-asyncio==1.3.0",
    "moto==5.1.22",
    "hypothesis==6.131.0",
]
security = [
    "bandit==1.9.4",
    "pip-tools==7.5.3",
]
dev = [
    "gco-cli[cdk,lint,typecheck,test,security]",
    "pre-commit==4.5.1",
]

[project.scripts]
gco = "cli.main:main"

[project.urls]
Documentation = "https://github.com/awslabs/global-capacity-orchestrator-on-aws/blob/main/docs/CLI.md"
Repository = "https://github.com/awslabs/global-capacity-orchestrator-on-aws"

# =============================================================================
# Tool Configurations
# =============================================================================

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313', 'py314']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | \.mypy_cache
    | \.pytest_cache
    | __pycache__
    | cdk\.out
    | dist
    | build
    | lambda/kubectl-applier-simple-build
    | lambda/helm-installer-build
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
skip = [".venv", "cdk.out", "dist", "build",
        "lambda/kubectl-applier-simple-build",
        "lambda/helm-installer-build"]
known_first_party = ["gco", "cli"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.ruff]
line-length = 100
target-version = "py314"
exclude = [
    ".git",
    ".venv",
    ".mypy_cache",
    ".pytest_cache",
    "__pycache__",
    "cdk.out",
    "dist",
    "build",
    "lambda/kubectl-applier-simple-build",
    "lambda/helm-installer-build",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ARG", "S101"]
"lambda/*" = ["ARG"]

[tool.mypy]
python_version = "3.14"
# --strict equivalent: every option enabled except "warn_unreachable" which has
# false positives on platform-conditional branches. Keep these in sync with CI.
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_any_generics = true
disallow_subclassing_any = true
no_implicit_optional = true
no_implicit_reexport = true
strict_equality = true
extra_checks = true
warn_redundant_casts = true
warn_unused_ignores = true
# Ignore missing imports from untyped third-party libraries (boto3, kubernetes,
# aws_cdk, etc. don't ship type stubs we can rely on). This is NOT a license
# to skip annotating our own code — it only suppresses errors from imports.
ignore_missing_imports = true
disable_error_code = ["import-untyped", "import-not-found"]
exclude = [
    "cdk\\.out",
    "\\.venv",
    "dist",
    "build",
    "lambda/kubectl-applier-simple-build",
    "lambda/helm-installer-build",
]

# Tests can use looser rules — the test runner resolves fixtures at runtime so
# `disallow_untyped_defs` creates too much noise for little real-world benefit.
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false
check_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["gco", "cli"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/cdk.out/*",
    "*/.venv/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
]
fail_under = 85
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.bandit]
exclude_dirs = ["tests", "cdk.out", ".venv"]
skips = ["B101", "B104", "B113", "B311"]
# B101: assert used in tests and CDK constructs (harmless)
# B104: bind to 0.0.0.0 required for container services
# B113: false positives on dict.get() calls named "requests" (all HTTP calls have timeouts)
# B311: random not used for security purposes
